refactor(api): group account-recovery routes into same folder
