+++ api/tests/identity-access-management/acceptance/application/account-recovery/account-recovery.route.test.js
@@ -1,7 +1,4 @@
-import sinon from 'sinon';
-
 import { createServer } from '../../../../../server.js';
-import { mailService } from '../../../../../src/shared/domain/services/mail-service.js';
 import { expect } from '../../../../test-helper.js';
 import { databaseBuilder, knex } from '../../../../tooling/databases.js';
 
@@ -118,12 +115,10 @@ describe('Acceptance | Identity Access Management | Application | Route | accoun
         const firstName = 'Gertrude';
         const username = 'gertrude.0202';
 
-        databaseBuilder.factory.buildUser.withRawPassword({ id: userId, username });
+        databaseBuilder.factory.buildUser.withRawPassword({ id: userId, username, firstName });
         databaseBuilder.factory.buildAuthenticationMethod.withGarAsIdentityProvider({ userId });
-        const { id: organizationLearnerId } = databaseBuilder.factory.buildOrganizationLearner({ userId, firstName });
         const { id } = databaseBuilder.factory.buildAccountRecoveryDemand({
           userId,
-          organizationLearnerId,
           temporaryKey,
           newEmail,
           used: false,
@@ -146,112 +141,6 @@ describe('Acceptance | Identity Access Management | Application | Route | accoun
         expect(response.result.data.attributes['has-gar-authentication-method']).to.be.true;
         expect(response.result.data.attributes['has-sco-username']).to.be.true;
       });
-    });
-  });
-
-  describe('POST /api/account-recovery', function () {
-    const studentInformation = {
-      ineIna: '123456789aa',
-      firstName: 'Jude',
-      lastName: 'Law',
-      birthdate: '2016-06-01',
-    };
-
-    const createUserWithSeveralOrganizationLearners = async ({ email = 'jude.law@example.net' } = {}) => {
-      const user = databaseBuilder.factory.buildUser.withRawPassword({
-        id: 8,
-        firstName: 'Judy',
-        lastName: 'Howl',
-        email,
-        username: 'jude.law0601',
-      });
-      const organization = databaseBuilder.factory.buildOrganization({
-        id: 7,
-        name: 'Collège Hollywoodien',
-      });
-      const latestOrganization = databaseBuilder.factory.buildOrganization({
-        id: 2,
-        name: 'Super Collège Hollywoodien',
-      });
-      databaseBuilder.factory.buildOrganizationLearner({
-        userId: user.id,
-        ...studentInformation,
-        nationalStudentId: studentInformation.ineIna.toUpperCase(),
-        organizationId: organization.id,
-        updatedAt: new Date('2005-01-01T15:00:00Z'),
-      });
-      databaseBuilder.factory.buildOrganizationLearner({
-        userId: user.id,
-        ...studentInformation,
-        nationalStudentId: studentInformation.ineIna.toUpperCase(),
-        organizationId: latestOrganization.id,
-        updatedAt: new Date('2010-01-01T15:00:00Z'),
-      });
-      await databaseBuilder.commit();
-    };
-
-    it('returns 204 HTTP status code', async function () {
-      // given
-      const sendAccountRecoveryEmailSpy = sinon.spy(mailService, 'sendAccountRecoveryEmail');
-
-      await createUserWithSeveralOrganizationLearners();
-      const newEmail = 'new_email@example.net';
-
-      const options = {
-        method: 'POST',
-        url: '/api/account-recovery',
-        payload: {
-          data: {
-            attributes: {
-              'ine-ina': studentInformation.ineIna,
-              'first-name': studentInformation.firstName,
-              'last-name': studentInformation.lastName,
-              birthdate: studentInformation.birthdate,
-              email: newEmail,
-            },
-          },
-        },
-      };
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(204);
-      expect(sendAccountRecoveryEmailSpy).to.have.been.calledWithExactly({
-        firstName: 'Jude',
-        email: 'new_email@example.net',
-        temporaryKey: sinon.match.string,
-      });
-    });
-
-    it('returns 422 if email already exists', async function () {
-      // given
-      const newEmail = 'new_email@example.net';
-      await createUserWithSeveralOrganizationLearners({ email: newEmail });
-
-      const options = {
-        method: 'POST',
-        url: '/api/account-recovery',
-        payload: {
-          data: {
-            attributes: {
-              'ine-ina': studentInformation.ineIna,
-              'first-name': studentInformation.firstName,
-              'last-name': studentInformation.lastName,
-              birthdate: studentInformation.birthdate,
-              email: newEmail,
-            },
-          },
-        },
-      };
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(422);
-      expect(response.result.errors[0].detail).to.equal('Invalid or already used e-mail address');
     });
   });
 });
